- replaced a large batch of DWORDs.
Most of those which still rely on ZDoom's own definition should be gone, unfortunately the code in files that include Windows headers is a gigantic mess with DWORDs being longs there intead of ints, so this needs to be done with care. DWORD should only remain where the Windows type is actually wanted.

# Conflicts:
#	src/f_wipe.cpp
#	src/gl/data/gl_data.cpp
#	src/gl/data/gl_portaldata.cpp
#	src/gl/data/gl_setup.cpp
#	src/gl/scene/gl_bsp.cpp
#	src/gl/scene/gl_renderhacks.cpp
#	src/gl/scene/gl_sprite.cpp
#	src/gl/system/gl_wipe.cpp
#	src/gl/textures/gl_material.cpp
#	src/gl/textures/gl_texture.cpp
#	src/gl/utility/gl_clock.cpp
#	src/r_data/colormaps.h
#	src/swrenderer/r_swcanvas.cpp
#	src/swrenderer/scene/r_portal.cpp
#	src/textures/warptexture.cpp