- implemented proper scaling for the status bar itself.
This allows using the UI scale or its own value, like all other scaling values.
In addition there is a choice between preserving equal pixel size or aspect ratio because the squashed non-corrected versions tend to look odd, but since proper scaling requires ununiform pixel sizes it is an option.

- changed how status bar sizes are being handled.

This has to recalculate all scaling and positioning factors, which can cause problems if the drawer leaves with some temporary values that do not reflect the status bar as a whole.
Changed it so that the status bar stores the base values and restores them after drawing is complete.

# Conflicts:
#	src/g_statusbar/shared_sbar.cpp
#	src/gl/scene/gl_weapon.cpp
#	src/polyrenderer/scene/poly_playersprite.cpp
#	src/r_utility.cpp
#	src/swrenderer/things/r_playersprite.cpp
#	src/v_video.cpp